requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
docling==2.5.2
sentence-transformers==2.2.2
openai==1.12.0
cohere==4.45
upstash-vector==0.3.1
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
